+++ back/app/services/form_logic_service.rb
@@ -29,6 +29,8 @@ class FormLogicService
 
       field_valid = if field.page?
         valid_page_logic_structure?(field) && valid_next_page?(field)
+      elsif field.section?
+        valid_section_logic_structure?(field)
       else
         valid_field_logic_structure?(field) && valid_rules?(field) && required?(field)
       end
@@ -56,6 +58,13 @@ class FormLogicService
       return true if all_rules_are_valid
     end
     add_invalid_structure_error(field)
+    false
+  end
+
+  def valid_section_logic_structure?(field)
+    return true unless field.logic
+
+    add_not_allowed_on_section_error(field)
     false
   end
 
@@ -116,6 +125,10 @@ class FormLogicService
 
     add_only_page_allowed_as_target_error(field, target_id)
     false
+  end
+
+  def add_not_allowed_on_section_error(field)
+    field.errors.add(:logic, :not_allowed_on_section_fields, message: 'not allowed on section fields')
   end
 
   def add_not_required_error(field)