+++ front/app/components/HookForm/InputMultilocWithLocaleSwitcher/index.tsx
@@ -7,7 +7,7 @@ import {
 } from '@citizenlab/cl2-component-library';
 import Error, { TFieldName } from 'components/UI/Error';
 import { Controller, useFormContext, FieldError } from 'react-hook-form';
-import { CLError, Locale } from 'typings';
+import { CLError, Locale, RHFErrors } from 'typings';
 import { get } from 'lodash-es';
 
 interface Props
@@ -35,14 +35,14 @@ const InputMultilocWithLocaleSwitcher = ({ name, ...rest }: Props) => {
   );
 
   // Select the first error messages from the field's multiloc validation error
-  const errors = get(formContextErrors, name);
+  const errors = get(formContextErrors, name) as RHFErrors;
+
   const validationError = Object.values(
     (errors as Record<Locale, FieldError> | undefined) || {}
   )[0]?.message;
 
   // If an API error with a matching name has been returned from the API response, apiError is set to an array with the error message as the only item
-  const apiError =
-    (errors?.error as string | undefined) && ([errors] as unknown as CLError[]);
+  const apiError = errors?.error && ([errors] as CLError[]);
 
   return (
     <>