+++ front/app/containers/Admin/projects/project/AdminProjectsProjectIndex.test.tsx
@@ -279,6 +279,7 @@ describe('<AdminProjectEdition />', () => {
       'General',
       'Description',
       'Input manager',
+      'Input form',
       'Allowed input tags',
       'Timeline',
       'Events',