+++ front/app/modules/commercial/insights/admin/containers/Insights/Details/Network/index.tsx
@@ -15,14 +15,11 @@ import ForceGraph2D, {
 import { forceCollide } from 'd3-force';
 
 // hooks
-import useInsightsView from 'modules/commercial/insights/hooks/useInsightsView';
-import useNetwork from 'modules/commercial/insights/hooks/useInsightsNetwork';
-
-// types
-import { IInsightsNetworkNode } from 'modules/commercial/insights/services/insightsNetwork';
+import useView from 'modules/commercial/insights/api/views/useView';
+import useNetwork from 'modules/commercial/insights/api/network/useNetwork';
 
 // utils
-import { isNilOrError, isError } from 'utils/helperUtils';
+import { isNilOrError } from 'utils/helperUtils';
 import { cloneDeep } from 'lodash-es';
 import { colors } from 'utils/styleUtils';
 import clHistory from 'utils/cl-router/history';
@@ -48,6 +45,7 @@ import { WrappedComponentProps } from 'react-intl';
 import messages from '../../messages';
 // styles
 import styled from 'styled-components';
+import { IInsightsNetworkNode } from 'modules/commercial/insights/api/network/types';
 
 type CanvasCustomRenderMode = 'replace' | 'before' | 'after';
 type Node = NodeObject & IInsightsNetworkNode;
@@ -88,8 +86,8 @@ const Network = ({
   const [zoomLevel, setZoomLevel] = useState(0);
 
   const networkRef = useRef<ForceGraphMethods>();
-  const { loading, network } = useNetwork(viewId);
-  const view = useInsightsView(viewId);
+  const { data: network, isLoading, isError } = useNetwork(viewId);
+  const { data: view } = useView(viewId);
 
   useEffect(() => {
     if (networkRef.current) {
@@ -228,7 +226,7 @@ const Network = ({
     }
   };
 
-  if (loading) {
+  if (isLoading) {
     return (
       <Centerer>
         <Spinner />
@@ -236,7 +234,7 @@ const Network = ({
     );
   }
 
-  if (isError(network)) {
+  if (isError) {
     return (
       <Box
         w="50%"
@@ -303,7 +301,7 @@ const Network = ({
           />
         </SectionTitle>
       </Box>
-      {height && width && (
+      {!!height && !!width && (
         <ForceGraph2D
           height={height}
           width={width}