+++ front/app/modules/commercial/insights/admin/containers/Insights/Edit/RenameCategory.tsx
@@ -14,13 +14,11 @@ import { Button, Input } from '@citizenlab/cl2-component-library';
 import Error from 'components/UI/Error';
 
 // services
-import { updateInsightsCategory } from 'modules/commercial/insights/services/insightsCategories';
+import useUpdateCategory from 'modules/commercial/insights/api/categories/useUpdateCategory';
 
 // utils
-import { isNilOrError } from 'utils/helperUtils';
 
 // types
-import { CLErrors } from 'typings';
 
 // tracking
 import { trackEventByName } from 'utils/analytics';
@@ -65,31 +63,23 @@ const RenameCategory = ({
   params: { viewId },
   location: { query },
 }: RenameCategoryProps) => {
-  const [loading, setLoading] = useState(false);
-  const [errors, setErrors] = useState<CLErrors | undefined>();
+  const { mutate, error, isLoading, reset } = useUpdateCategory();
 
   const [name, setName] = useState<string>(originalCategoryName);
+
   const onChangeName = (value: string) => {
     setName(value);
-    setErrors(undefined);
+    reset();
   };
 
-  const handleSubmit = async () => {
+  const handleSubmit = () => {
     if (name) {
-      setLoading(true);
-      try {
-        const result = await updateInsightsCategory(
-          viewId,
-          query.category,
-          name
-        );
-        if (!isNilOrError(result)) {
-          closeRenameModal();
+      mutate(
+        { viewId, categoryId: query.category, requestBody: { name } },
+        {
+          onSuccess: closeRenameModal,
         }
-      } catch (errors) {
-        setErrors(errors.json.errors);
-        setLoading(false);
-      }
+      );
       trackEventByName(tracks.editCategory);
     }
   };
@@ -104,13 +94,13 @@ const RenameCategory = ({
           onChange={onChangeName}
           label={formatMessage(messages.renameCategoryModalNameLabel)}
         />
-        {errors && (
-          <Error apiErrors={errors['name']} fieldName="category_name" />
+        {error && (
+          <Error apiErrors={error.errors['name']} fieldName="category_name" />
         )}
 
         <ButtonContainer>
           <Button
-            processing={loading}
+            processing={isLoading}
             disabled={!name}
             onClick={handleSubmit}
             bgColor={colors.primary}