+++ front/app/modules/commercial/insights/api/inputs/useInfiniteInputs.test.ts
@@ -1,0 +1,100 @@
+import { renderHook } from '@testing-library/react-hooks';
+
+import useInfiniteInputs from './useInfiniteInputs';
+import views from '../../fixtures/views';
+
+import { setupServer } from 'msw/node';
+import { rest } from 'msw';
+
+import createQueryClientWrapper from 'utils/testUtils/queryClientWrapper';
+import links from '../../fixtures/links';
+
+const apiPath = '*insights/views/:viewId/inputs';
+const server = setupServer(
+  rest.get(apiPath, (_req, res, ctx) => {
+    return res(ctx.status(200), ctx.json({ data: views, links }));
+  })
+);
+
+describe('useInfiniteInputs', () => {
+  beforeAll(() => server.listen());
+  afterAll(() => server.close());
+
+  it('returns data correctly with next page', async () => {
+    const { result, waitFor } = renderHook(
+      () =>
+        useInfiniteInputs('viewId', {
+          categories: [],
+          keywords: [],
+          pageNumber: 1,
+          search: '',
+        }),
+      {
+        wrapper: createQueryClientWrapper(),
+      }
+    );
+
+    expect(result.current.isLoading).toBe(true);
+
+    await waitFor(() => expect(result.current.isSuccess).toBe(true));
+
+    expect(result.current.isLoading).toBe(false);
+    expect(result.current.data?.pages[0].data).toEqual(views);
+    expect(result.current.hasNextPage).toBe(true);
+  });
+
+  it('returns data correctly with no next page', async () => {
+    const newLinks = { ...links, next: null };
+    server.use(
+      rest.get(apiPath, (_req, res, ctx) => {
+        return res(ctx.status(200), ctx.json({ data: views, links: newLinks }));
+      })
+    );
+    const { result, waitFor } = renderHook(
+      () =>
+        useInfiniteInputs('viewId', {
+          categories: [],
+          keywords: [],
+          pageNumber: 1,
+          search: '',
+        }),
+      {
+        wrapper: createQueryClientWrapper(),
+      }
+    );
+
+    expect(result.current.isLoading).toBe(true);
+
+    await waitFor(() => expect(result.current.isSuccess).toBe(true));
+
+    expect(result.current.isLoading).toBe(false);
+    expect(result.current.data?.pages[0].data).toEqual(views);
+    expect(result.current.hasNextPage).toBe(false);
+  });
+
+  it('returns error correctly', async () => {
+    server.use(
+      rest.get(apiPath, (_req, res, ctx) => {
+        return res(ctx.status(500));
+      })
+    );
+
+    const { result, waitFor } = renderHook(
+      () =>
+        useInfiniteInputs('viewId', {
+          categories: [],
+          keywords: [],
+          pageNumber: 1,
+          search: '',
+        }),
+      {
+        wrapper: createQueryClientWrapper(),
+      }
+    );
+
+    expect(result.current.isLoading).toBe(true);
+    await waitFor(() => expect(result.current.isError).toBe(true));
+    expect(result.current.error).toBeDefined();
+    expect(result.current.isLoading).toBe(false);
+  });
+});