+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 2023-02-15 (2)
+
+### Added
+
+- [CL-1895] Flexible input form - form builder now in use for editing ideation forms with additional custom fields
+
+### Fixed
+
+- [CL-2830] Fixed bug that prevented editing ideas
+
+## 2023-02-15
+
+### Added
+
+- [CL-2836] Admin shortcut from home page for users with admin access
+
 ## 2023-02-13 (2)
 
 ### Changed